Move API handlers to separate package
This makes the package split a little cleaner and easier to import the
HTTP handlers for other consumers.
